move collectionish tile code to .tiles
