+++ src/collective/collectionfilter/configure.zcml
@@ -49,13 +49,8 @@
       />
 
   <adapter factory=".filteritems.CollectionishCollection"
-           provides=".filteritems.ICollectionish"
+           provides="collective.collectionfilter.interfaces.ICollectionish"
            for="plone.app.contenttypes.interfaces.ICollection" />
-
-  <adapter factory=".filteritems.CollectionishLayout"
-           provides=".filteritems.ICollectionish"
-           for="plone.app.blocks.layoutbehavior.ILayoutBehaviorAdaptable" 
-           zcml:condition="installed plone.tiles" />
 
   <subscriber
       for="Products.CMFPlone.interfaces.IPloneSiteRoot
