+++ src/collective/collectionfilter/filteritems.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from collective.collectionfilter import _
 from collective.collectionfilter.interfaces import IGroupByCriteria
+from collective.collectionfilter.interfaces import ICollectionish
 from collective.collectionfilter.query import make_query
 from collective.collectionfilter.utils import base_query
 from collective.collectionfilter.utils import safe_decode
@@ -9,12 +10,7 @@ from collective.collectionfilter.utils import safe_iterable
 from collective.collectionfilter.vocabularies import DEFAULT_FILTER_TYPE
 from collective.collectionfilter.vocabularies import EMPTY_MARKER
 from Missing import Missing
-from plone.app.blocks.layoutbehavior import ILayoutBehaviorAdaptable
 from plone.app.contenttypes.behaviors.collection import ICollection
-try:
-    from plone.app.blocks.layoutbehavior import ILayoutAware
-except ImportError:
-    ILayoutAware = None
 from plone.app.event.base import _prepare_range
 from plone.app.event.base import guess_date_from
 from plone.app.event.base import start_end_from_mode
@@ -24,17 +20,13 @@ from plone.i18n.normalizer import idnormalizer
 from plone.memoize import ram
 from plone.memoize.volatile import DontCache
 from six.moves.urllib.parse import urlencode
-from zope.component import adapter
 from zope.component import getUtility
-from zope.component import getMultiAdapter
-from zope.interface import Interface
 from zope.interface import implementer
 from zope.globalrequest import getRequest
 from zope.i18n import translate
 
 import plone.api
 import six
-import re
 
 
 try:
@@ -244,10 +236,6 @@ def get_filter_items(
     return ret
 
 
-class ICollectionish(Interface):
-    "Adapts object similar to ICollection if has contentlisting tile, or if collection"
-
-
 @implementer(ICollectionish)
 class CollectionishCollection(object):
 
@@ -295,73 +283,3 @@ class CollectionishCollection(object):
             #        do that
 
         return self.collection.results(batch=False, brains=True, custom_query=custom_query)
-
-
-if ILayoutAware is not None:
-    @implementer(ICollectionish)
-    @adapter(ILayoutBehaviorAdaptable)
-    class CollectionishLayout(CollectionishCollection):
-        """Provide interface for either objects with contentlisting tiles or collections or both"""
-
-        tile = None
-
-        def __init__(self, context):
-            self.context = context
-
-            la = ILayoutAware(self.context)
-            if la.content:
-                urls = re.findall('(@@plone.app.standardtiles.contentlisting/[^"]+)', la.content)
-                if urls:
-                    # TODO: maybe better to get tile data? using ITileDataManager(id)?
-                    url = context.REQUEST.response.headers.get('x-tile-url')
-                    tile = self.context.unrestrictedTraverse(urls[0])
-                    tile.update()
-                    if context.REQUEST.response.headers.get('x-tile-url'):
-                        if url:
-                            context.REQUEST.response.headers['x-tile-url'] = url
-                        else:
-                            del context.REQUEST.response.headers['x-tile-url']
-
-                    # print(context.REQUEST.response.headers)
-                    self.tile = tile
-            if self.tile is None:
-                # Could still be a ILayoutAware collection
-                try:
-                    self.collection = ICollection(self.context)
-                except TypeError:
-                    raise TypeError("No contentlisting tile or Collection found")
-            else:
-                self.collection = self.tile  # to get properties
-
-        @property
-        def sort_reversed(self):
-            if self.tile is not None:
-                return self.sort_order == "reverse"
-            else:
-                return self.collection.sort_reversed
-
-        def results(self, custom_query, request_params):
-            """Search results"""
-            if self.tile is None:
-                return super(CollectionishLayout, self).results(custom_query, request_params)
-
-            builder = getMultiAdapter(
-                (self.context, self.context.REQUEST), name="querybuilderresults"
-            )
-
-            # Include query parameters from request if not set to ignore
-            contentFilter = {}
-            if not getattr(self.tile, "ignore_request_params", False):
-                contentFilter = dict(self.context.REQUEST.get("contentFilter", {}))
-
-            # TODO: handle events extra params
-
-            return builder(
-                query=self.query,
-                sort_on=self.sort_on or "getObjPositionInParent",
-                sort_order=self.sort_order,
-                limit=self.limit,
-                batch=False,
-                brains=True,
-                custom_query=custom_query if custom_query is not None else contentFilter,
-            )
