+++ src/collective/collectionfilter/interfaces.py
@@ -33,6 +33,10 @@ def pattern_options():
     return options
 
 
+class ICollectionish(Interface):
+    "Adapts object similar to ICollection if has contentlisting tile, or is a collection"
+
+
 class ICollectionFilterBaseSchema(Interface):
 
     header = schema.TextLine(
@@ -45,7 +49,8 @@ class ICollectionFilterBaseSchema(Interface):
         title=_(u"label_target_collection", default=u"Alternative Target Collection"),
         description=_(
             u"help_target_collection",
-            default=u"We use the current context as collection. As an alternative you can select a different collection as source for the filter items "
+            default=u"We use the current context as collection. As an alternative you can select a different collection "
+            u"as source for the filter items "
             u"and where the filter is applied.",
         ),
         required=False,
