+++ src/collective/collectionfilter/tiles/configure.zcml
@@ -8,6 +8,12 @@
     <include package="plone.tiles" />
     <include package="plone.tiles" file="meta.zcml"/>
 
+
+  <adapter factory=".CollectionishLayout"
+           provides="collective.collectionfilter.interfaces.ICollectionish"
+           for="plone.app.blocks.layoutbehavior.ILayoutBehaviorAdaptable" 
+           />
+
     <plone:tile
         name="collective.collectionfilter.tiles.filter"
         title="Collection Filter Tile"
